package/dbusbroker: new package
Add dbus-broker, which is a drop-in replacement
for the dbus-daemon.

Its possible to use this package standalone (without the dbus
package - if buildroot's systemd would not depend on dbus).
This is sufficient to provide systemd's (d)bus functionality.
To allow standalone usage, the necessary config files are
copied and adopted over from dbus.

bases on Yanns changes, and

-   add an own config entry for dbus-broker-launch
    enabled by default if systemd init is used
-   undo BR2_COREUTILS_HOST_DEPENDENCY
-   undo adding dbus user - never used by this package
-   add condtional audit dependency
-   cleanup conditional logic a bit

Signed-off-by: Norbert Lange <nolange79@gmail.com>